+++ src/integration-test/java/com/commercetools/sync/integration/ctpprojectsource/cartdiscounts/CartDiscountSyncIT.java
@@ -14,9 +14,10 @@ import io.sphere.sdk.cartdiscounts.ShippingCostTarget;
 import io.sphere.sdk.cartdiscounts.commands.updateactions.ChangeCartPredicate;
 import io.sphere.sdk.cartdiscounts.commands.updateactions.ChangeTarget;
 import io.sphere.sdk.cartdiscounts.commands.updateactions.ChangeValue;
-import io.sphere.sdk.cartdiscounts.queries.CartDiscountQuery;
+import io.sphere.sdk.cartdiscounts.commands.updateactions.SetCustomType;
 import io.sphere.sdk.commands.UpdateAction;
 import io.sphere.sdk.types.CustomFieldsDraft;
+import io.sphere.sdk.types.Type;
 import io.sphere.sdk.utils.MoneyImpl;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -25,18 +26,21 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 import java.util.stream.Collectors;
 
+import static com.commercetools.sync.cartdiscounts.utils.CartDiscountReferenceReplacementUtils.buildCartDiscountQuery;
+import static com.commercetools.sync.cartdiscounts.utils.CartDiscountReferenceReplacementUtils.replaceCartDiscountsReferenceIdsWithKeys;
 import static com.commercetools.sync.commons.asserts.statistics.AssertionsForStatistics.assertThat;
-import static com.commercetools.sync.integration.commons.utils.CartDiscountITUtils.OLD_CART_DISCOUNT_TYPE_KEY;
+import static com.commercetools.sync.integration.commons.utils.CartDiscountITUtils.createCartDiscountCustomType;
 import static com.commercetools.sync.integration.commons.utils.CartDiscountITUtils.deleteCartDiscountsFromTargetAndSource;
 import static com.commercetools.sync.integration.commons.utils.CartDiscountITUtils.populateSourceProject;
 import static com.commercetools.sync.integration.commons.utils.CartDiscountITUtils.populateTargetProject;
-import static com.commercetools.sync.integration.commons.utils.ITUtils.createCustomFieldsJsonMap;
 import static com.commercetools.sync.integration.commons.utils.ITUtils.deleteTypesFromTargetAndSource;
 import static com.commercetools.sync.integration.commons.utils.SphereClientUtils.CTP_SOURCE_CLIENT;
 import static com.commercetools.sync.integration.commons.utils.SphereClientUtils.CTP_TARGET_CLIENT;
 import static io.sphere.sdk.models.DefaultCurrencyUnits.EUR;
+import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class CartDiscountSyncIT {
@@ -59,27 +63,10 @@ class CartDiscountSyncIT {
     void sync_WithoutUpdates_ShouldReturnProperStatistics() {
         // preparation
         final List<CartDiscount> cartDiscounts = CTP_SOURCE_CLIENT
-            .execute(CartDiscountQuery.of())
+            .execute(buildCartDiscountQuery())
             .toCompletableFuture().join().getResults();
 
-        final List<CartDiscountDraft> cartDiscountDrafts = cartDiscounts
-            .stream()
-            .map(cartDiscount ->
-                CartDiscountDraftBuilder.of(cartDiscount.getName(),
-                    cartDiscount.getCartPredicate(),
-                    cartDiscount.getValue(),
-                    cartDiscount.getTarget(),
-                    cartDiscount.getSortOrder(),
-                    cartDiscount.isRequiringDiscountCode())
-                                        .key(cartDiscount.getKey())
-                                        .active(cartDiscount.isActive())
-                                        .description(cartDiscount.getDescription())
-                                        .validFrom(cartDiscount.getValidFrom())
-                                        .validUntil(cartDiscount.getValidUntil())
-                                        .custom(CustomFieldsDraft
-                                            .ofTypeIdAndJson(OLD_CART_DISCOUNT_TYPE_KEY, createCustomFieldsJsonMap()))
-                                        .build())
-            .collect(Collectors.toList());
+        final List<CartDiscountDraft> cartDiscountDrafts = replaceCartDiscountsReferenceIdsWithKeys(cartDiscounts);
 
         final List<String> errorMessages = new ArrayList<>();
         final List<Throwable> exceptions = new ArrayList<>();
@@ -114,28 +101,27 @@ class CartDiscountSyncIT {
     void sync_WithUpdates_ShouldReturnProperStatistics() {
         // preparation
         final List<CartDiscount> cartDiscounts = CTP_SOURCE_CLIENT
-            .execute(CartDiscountQuery.of())
+            .execute(buildCartDiscountQuery())
             .toCompletableFuture().join().getResults();
+        final String newTypeKey = "new-type";
+        createCartDiscountCustomType(newTypeKey, Locale.ENGLISH, newTypeKey, CTP_SOURCE_CLIENT);
+        final Type newTargetCustomType = createCartDiscountCustomType(newTypeKey, Locale.ENGLISH, newTypeKey,
+            CTP_TARGET_CLIENT);
 
-        final List<CartDiscountDraft> cartDiscountDrafts = cartDiscounts
+        final List<CartDiscountDraft> cartDiscountsReferenceIdsWithKeys =
+            replaceCartDiscountsReferenceIdsWithKeys(cartDiscounts);
+
+        // Apply some changes
+        final List<CartDiscountDraft> cartDiscountDrafts = cartDiscountsReferenceIdsWithKeys
             .stream()
-            .map(cartDiscount ->
-                CartDiscountDraftBuilder.of(cartDiscount.getName(),
-                    CartPredicate.of("totalPrice >= \"100 EUR\""), //new cart predicate
-                    AbsoluteCartDiscountValue.of(MoneyImpl.of(40, EUR)), //new value
-                    ShippingCostTarget.of(), //new target
-                    cartDiscount.getSortOrder(),
-                    cartDiscount.isRequiringDiscountCode())
-                                        .key(cartDiscount.getKey())
-                                        .active(cartDiscount.isActive())
-                                        .description(cartDiscount.getDescription())
-                                        .validFrom(cartDiscount.getValidFrom())
-                                        .validUntil(cartDiscount.getValidUntil())
-                                        .custom(CustomFieldsDraft
-                                            .ofTypeIdAndJson(OLD_CART_DISCOUNT_TYPE_KEY, createCustomFieldsJsonMap()))
-                                        .build())
+            .map(draft -> CartDiscountDraftBuilder
+                .of(draft)
+                .cartPredicate(CartPredicate.of("totalPrice >= \"100 EUR\""))
+                .value(AbsoluteCartDiscountValue.of(MoneyImpl.of(40, EUR)))
+                .target(ShippingCostTarget.of())
+                .custom(CustomFieldsDraft.ofTypeIdAndJson(newTypeKey, emptyMap()))
+                .build())
             .collect(Collectors.toList());
-
 
         final List<String> errorMessages = new ArrayList<>();
         final List<Throwable> exceptions = new ArrayList<>();
@@ -164,9 +150,10 @@ class CartDiscountSyncIT {
         assertThat(errorMessages).isEmpty();
         assertThat(exceptions).isEmpty();
         assertThat(updateActionsList).containsExactly(
-                ChangeValue.of(CartDiscountValue.ofAbsolute(Collections.singletonList(MoneyImpl.of(40, EUR)))),
-                ChangeCartPredicate.of("totalPrice >= \"100 EUR\""),
-                ChangeTarget.of(ShippingCostTarget.of())
+            ChangeValue.of(CartDiscountValue.ofAbsolute(Collections.singletonList(MoneyImpl.of(40, EUR)))),
+            ChangeCartPredicate.of("totalPrice >= \"100 EUR\""),
+            ChangeTarget.of(ShippingCostTarget.of()),
+            SetCustomType.ofTypeIdAndJson(newTargetCustomType.getId(), emptyMap())
         );
         assertThat(cartDiscountSyncStatistics).hasValues(2, 1, 1, 0);
         assertThat(cartDiscountSyncStatistics
@@ -174,5 +161,4 @@ class CartDiscountSyncIT {
             .isEqualTo("Summary: 2 cart discounts were processed in total"
                 + " (1 created, 1 updated and 0 failed to sync).");
     }
-
 }