+++ docs/RELEASE_NOTES.md
@@ -29,15 +29,25 @@
 -->
 
 <!--
-### 1.3.0 -  Jun 24, 2019
+### 1.4.0 -  Jun 24, 2019
+[Commits](https://github.com/commercetools/commercetools-sync-java/compare/1.3.0...1.4.0) |
+[Javadoc](https://commercetools.github.io/commercetools-sync-java/v/1.4.0/) | 
+[Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/1.4.0)
+
+-->
+
+### 1.3.0 -  Jul 3, 2019
 [Commits](https://github.com/commercetools/commercetools-sync-java/compare/1.2.0...1.3.0) |
 [Javadoc](https://commercetools.github.io/commercetools-sync-java/v/1.3.0/) | 
 [Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/1.3.0)
 
-- 🎉 **New Features** (3)
+- 🎉 **New Features** (6)
     - **CartDiscount Sync** - Added support for syncing types. [#379](https://github.com/commercetools/commercetools-sync-java/issues/379) For more info how to use it please refer to [CartDiscount usage doc](/docs/usage/CART_DISCOUNT_SYNC.md).
     - **CartDiscount Sync** - Exposed `CartDiscountSyncUtils#buildActions` which calculates all needed update actions after comparing a `CartDiscount` and a `CartDiscountDraft`. [[#379](https://github.com/commercetools/commercetools-sync-java/issues/379)
     - **CartDiscount Sync** - Exposed `CartDiscountUpdateActionUtils` which contains utils for calculating needed update actions after comparing individual fields of a `CartDiscount` and a `CartDiscountDraft`. [#379](https://github.com/commercetools/commercetools-sync-java/issues/379)
+    - **CartDiscount Sync** - Introduced the new `CartDiscountReferenceResolver` which resolves custom type references on CartDiscountDrafts. [#379](https://github.com/commercetools/commercetools-sync-java/issues/379)
+    - **CartDiscount Sync** - Introduced new `CartDiscountReferenceReplacementUtils#replaceCartDiscountsReferenceIdsWithKeys` which is a util that replaces the custom type ids with keys in a list of cartDiscounts. [#379](https://github.com/commercetools/commercetools-sync-java/issues/379)
+    - **CartDiscount Sync** - Exposed `CartDiscountReferenceReplacementUtils#buildCartDiscountQuery` util to create a cart discount query with all needed reference expansions to fetch cart discounts from a source CTP project for the sync. [#379](https://github.com/commercetools/commercetools-sync-java/issues/379).
     
 - 🐞 **Bug Fixes** (1)
     - **Commons** -  Fixed a bug in the `BaseSyncStatistics` which caused a wrong calculation of the `latestBatchProcessingTimeInMinutes`. [#378](https://github.com/commercetools/commercetools-sync-java/issues/378)
@@ -49,8 +59,6 @@
     - **Commons** - Bumped `org.junit.jupiter:junit-jupiter-api` to 5.5.0.
     - **Commons** - Bumped `org.junit.jupiter:junit-jupiter-engine` to 5.5.0.
     - **Commons** - Bumped `org.junit.jupiter:junit-jupiter-params` to 5.5.0.
-
--->
 
 ### 1.2.0 -  Jun 14, 2019
 [Commits](https://github.com/commercetools/commercetools-sync-java/compare/1.1.1...1.2.0) |